refactor: isolate sidebar input handling
